ARQ-322 Support for GF 3.1.2
Added support for GlassFish 3.1.2 via the use of the CsrfProtectionFilter in
the REST client. This allows deployments to occur against a GlassFish 3.1.2 instance, since GF 3.1.2 has a server-side CSRF protection filter. A corresponding
client-side protection filter is required for REST clients.

Deleted unnecessary files from the managed GlassFish adapter.

Added the allowConnectingToRunningServer property.

This allows Arquillian to now connect to a running GlassFish DAS, even
if Arquillian did not start the DAS.

Modified the build lifecycle to use the Maven Dependency plugin.

This now allows the GlassFish distribution to be downloaded off the
Maven central repository only once, instead of the GlassFish site
everytime, after a mvn clean is issued.

Added the arquillian-testenricher-initialcontext dependency.

Turns out the CDI injection wasnt working because of this missing dep.
Added the remaining tests from the glassfish-remote module, to ensure
the managed container is as good as the remote container impl.