ARQ-1232 Fixing random "contextRoot must not be null" failures
Performs a lookup of all the sub-components in a deployment. This
bypasses the GlassFish RESTful Admin bug since it no longer filters on
servlets alone. The existing codebase in the adapter automatically
discards metadata of components like EJB JARs or StatelessSessionBeans
since it is programmed to look only for WebModules and Servlets to
prepare the HTTPContext.